# index.js
// const express = require('express');
// const app = express();
// const port = 3000; // You can choose any available port

// app.get('/', (req, res) => {
//   res.send('Hello From Below! And Above');
// });

// app.listen(port, () => {
//   console.log(`Server is running on http://localhost:${port}`);
// });





// const express = require('express');
// const swaggerUi = require('swagger-ui-express');
// const swaggerSpec = require('./swagger'); // Import the generated Swagger documentation

// const app = express();
// const port = 3000;

// app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));

// app.get('/', (req, res) => {
//   res.send('Hello World!');
// });

// app.listen(port, () => {
//   console.log(`Server is running on http://localhost:${port}`);
// });


const express = require('express');
const swaggerUi = require('swagger-ui-express');
const swaggerFile = require('./swagger_output.json'); // Generated by swagger-autogen

const app = express();
const port = 3000;

// Enable JSON parsing middleware
app.use(express.json());

// Swagger UI setup
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerFile));

// Your routes
app.get('/TestGet', (req, res) => {
  res.send('Hello World!');
});

app.post('/TestPost', (req, res) => {
  const requestData = req.body.data || 'Default String';
  res.status(201).send(`Received data: ${requestData}`);
});

app.post('/SayHello', (req, res) => {
    const requestData = req.body.data || 'Default String';
    res.status(201).send(`Hello, ${requestData}`);
  });


app.post('/process-data', (req, res) => {
const { name, age, ethnicity, height } = req.body;

if (!name || !age || !ethnicity) {
    return res.status(400).json({ error: 'Missing required fields: name, age, ethnicity' });
}

const formattedResponse = {
    "Name": name || 'No name provided',
    "Age": age || 'No age provided',
    "Ethnicity": ethnicity ||  'No ethnicity provided',
    "Height": height || 'No height provided'
  };

res.json(formattedResponse);
});

app.listen(port, () => {
  console.log(`Server is running on http://localhost:${port}`);
});